+++ src/cli/backup/sharepoint.go
@@ -110,6 +110,7 @@ func addSharePointCommands(cmd *cobra.Command) *cobra.Command {
 
 		addFailedItemsFN(c)
 		addSkippedItemsFN(c)
+		addRecoveredErrorsFN(c)
 
 	case detailsCommand:
 		c, fs = utils.AddCommand(cmd, sharePointDetailsCmd())