+++ src/pkg/fault/fault_test.go
@@ -6,11 +6,11 @@ import (
 	"fmt"
 	"testing"
 
+	"github.com/alcionai/clues"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 	"github.com/stretchr/testify/suite"
 
-	"github.com/alcionai/clues"
 	"github.com/alcionai/corso/src/internal/tester"
 	"github.com/alcionai/corso/src/pkg/fault"
 )
@@ -211,8 +211,8 @@ func (suite *FaultErrorsUnitSuite) TestErrors() {
 	n.AddRecoverable(errors.New("2"))
 
 	d := n.Errors()
-	assert.Equal(t, n.Failure(), d.Failure)
-	assert.ElementsMatch(t, n.Recovered(), d.Recovered)
+	assert.Equal(t, clues.ToCore(n.Failure()), d.Failure)
+	assert.Len(t, d.Recovered, len(n.Recovered()))
 	assert.False(t, d.FailFast)
 
 	// fail-fast
@@ -224,90 +224,98 @@ func (suite *FaultErrorsUnitSuite) TestErrors() {
 	n.AddRecoverable(errors.New("2"))
 
 	d = n.Errors()
-	assert.Equal(t, n.Failure(), d.Failure)
-	assert.ElementsMatch(t, n.Recovered(), d.Recovered)
+	assert.Equal(t, clues.ToCore(n.Failure()), d.Failure)
+	assert.Len(t, d.Recovered, len(n.Recovered()))
 	assert.True(t, d.FailFast)
 }
 
 func (suite *FaultErrorsUnitSuite) TestErrors_Items() {
-	ae := assert.AnError
+	ae := clues.Stack(assert.AnError)
+	noncore := []*clues.ErrCore{ae.Core()}
 	addtl := map[string]any{"foo": "bar", "baz": 1}
 
 	table := []struct {
-		name   string
-		errs   func() fault.Errors
-		expect []fault.Item
+		name              string
+		errs              func() *fault.Errors
+		expectItems       []fault.Item
+		expectRecoverable []*clues.ErrCore
 	}{
 		{
 			name: "no errors",
-			errs: func() fault.Errors {
+			errs: func() *fault.Errors {
 				return fault.New(false).Errors()
 			},
-			expect: []fault.Item{},
+			expectItems:       []fault.Item{},
+			expectRecoverable: []*clues.ErrCore{},
 		},
 		{
 			name: "no items",
-			errs: func() fault.Errors {
+			errs: func() *fault.Errors {
 				b := fault.New(false)
 				b.Fail(ae)
 				b.AddRecoverable(ae)
 
 				return b.Errors()
 			},
-			expect: []fault.Item{},
+			expectItems:       []fault.Item{},
+			expectRecoverable: noncore,
 		},
 		{
 			name: "failure item",
-			errs: func() fault.Errors {
+			errs: func() *fault.Errors {
 				b := fault.New(false)
 				b.Fail(fault.OwnerErr(ae, "id", "name", addtl))
 				b.AddRecoverable(ae)
 
 				return b.Errors()
 			},
-			expect: []fault.Item{*fault.OwnerErr(ae, "id", "name", addtl)},
+			expectItems:       []fault.Item{*fault.OwnerErr(ae, "id", "name", addtl)},
+			expectRecoverable: noncore,
 		},
 		{
 			name: "recoverable item",
-			errs: func() fault.Errors {
+			errs: func() *fault.Errors {
 				b := fault.New(false)
 				b.Fail(ae)
 				b.AddRecoverable(fault.OwnerErr(ae, "id", "name", addtl))
 
 				return b.Errors()
 			},
-			expect: []fault.Item{*fault.OwnerErr(ae, "id", "name", addtl)},
+			expectItems:       []fault.Item{*fault.OwnerErr(ae, "id", "name", addtl)},
+			expectRecoverable: []*clues.ErrCore{},
 		},
 		{
 			name: "two items",
-			errs: func() fault.Errors {
+			errs: func() *fault.Errors {
 				b := fault.New(false)
 				b.Fail(fault.OwnerErr(ae, "oid", "name", addtl))
 				b.AddRecoverable(fault.FileErr(ae, "fid", "name", addtl))
 
 				return b.Errors()
 			},
-			expect: []fault.Item{
+			expectItems: []fault.Item{
 				*fault.OwnerErr(ae, "oid", "name", addtl),
 				*fault.FileErr(ae, "fid", "name", addtl),
 			},
+			expectRecoverable: []*clues.ErrCore{},
 		},
 		{
 			name: "duplicate items - failure priority",
-			errs: func() fault.Errors {
+			errs: func() *fault.Errors {
 				b := fault.New(false)
 				b.Fail(fault.OwnerErr(ae, "id", "name", addtl))
 				b.AddRecoverable(fault.FileErr(ae, "id", "name", addtl))
 
 				return b.Errors()
 			},
-			expect: []fault.Item{
+			expectItems: []fault.Item{
 				*fault.OwnerErr(ae, "id", "name", addtl),
 			},
+			expectRecoverable: []*clues.ErrCore{},
 		},
 		{
 			name: "duplicate items - last recoverable priority",
-			errs: func() fault.Errors {
+			errs: func() *fault.Errors {
 				b := fault.New(false)
 				b.Fail(ae)
 				b.AddRecoverable(fault.FileErr(ae, "fid", "name", addtl))
@@ -315,14 +323,41 @@ func (suite *FaultErrorsUnitSuite) TestErrors_Items() {
 
 				return b.Errors()
 			},
-			expect: []fault.Item{
+			expectItems: []fault.Item{
 				*fault.FileErr(ae, "fid", "name2", addtl),
 			},
+			expectRecoverable: []*clues.ErrCore{},
+		},
+		{
+			name: "recoverable item and non-items",
+			errs: func() *fault.Errors {
+				b := fault.New(false)
+				b.Fail(ae)
+				b.AddRecoverable(fault.FileErr(ae, "fid", "name", addtl))
+				b.AddRecoverable(ae)
+
+				return b.Errors()
+			},
+			expectItems: []fault.Item{
+				*fault.FileErr(ae, "fid", "name", addtl),
+			},
+			expectRecoverable: noncore,
 		},
 	}
 	for _, test := range table {
 		suite.Run(test.name, func() {
-			assert.ElementsMatch(suite.T(), test.expect, test.errs().Items)
+			t := suite.T()
+			fe := test.errs()
+
+			assert.ElementsMatch(t, test.expectItems, fe.Items)
+			require.Equal(t, test.expectRecoverable, fe.Recovered)
+
+			for i := range test.expectRecoverable {
+				expect := test.expectRecoverable[i]
+				got := fe.Recovered[i]
+
+				assert.Equal(t, *expect, *got)
+			}
 		})
 	}
 }