+++ src/pkg/fault/testdata/testdata.go
@@ -11,11 +11,11 @@ func MakeErrors(failure, recovered, skipped bool) fault.Errors {
 	fe := fault.Errors{}
 
 	if failure {
-		fe.Failure = assert.AnError
+		fe.Failure = clues.Wrap(assert.AnError, "wrapped").Core()
 	}
 
 	if recovered {
-		fe.Recovered = []error{clues.New("recoverable")}
+		fe.Recovered = []*clues.ErrCore{clues.New("recoverable").Core()}
 	}
 
 	if skipped {